WIP Create replace-secret script
The idea was to create a systemd service which is run on each config
rebuild. It would then replace the secrets in other config/service
files with values from /run/secrets.

But considering this will have to restart a bunch of services and handle
other runtime changes, it seems like a dead-end.

Another way this possibly could be solved is to create a fuse-based
program which takes a list of files and replace regexes. Then it exposes
a directory with the same files, but with replaced values. These files
are then bind-mounted over the original symlinks at their initial location.
As Nix uses symlinks to place config-files in the correct places, it would have to
be able to handle the re-creation of symlinks on rebuilds. Maybe Nix
fails if it can't find the expected symlink, but finds a file instead?

Maybe the bind-mount should happen before the symlink is even created.
Then as soon as it is created the program exposes a symlink to a file
with the values replaced with secrets instead?
When Nix then replaces the symlink with a new one (pointing to another
location in the /nix/store) it will read the new file, replace the
values with secrets and expose it, then point the symlink to it.